% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dostats.R
\name{class.stats}
\alias{class.stats}
\alias{numeric.stats}
\alias{factor.stats}
\alias{integer.stats}
\title{Filter by class}
\usage{
class.stats(.class)

numeric.stats(x, ...)

factor.stats(x, ...)

integer.stats(x, ...)
}
\arguments{
\item{.class}{string for class to filter by}

\item{x}{vector of any class}

\item{...}{passed to \code{\link{dostats}}}
}
\value{
data frame of computed statistics if x is of class \code{.class}
otherwise returns \code{NULL}.
}
\description{
Filter by class
}
\section{Functions}{
\itemize{
\item \code{numeric.stats}: Numeric class statistics

\item \code{factor.stats}: Factor class statistics

\item \code{integer.stats}: Integer class statistics
@export
}}

\seealso{
\code{\link{dostats}}
}
