% Generated by roxygen2 (4.0.0): do not edit by hand
\name{hdf}
\alias{hdf}
\title{Heirachical Data Frames}
\usage{
hdf(...)
}
\arguments{
\item{...}{vector, data.frame, or hdf objects to bind together.}
}
\description{
The primary purpose of the hdf is to represent data with grouped columns
and grouped rows.
}

