% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{first}
\alias{first}
\alias{last}
\title{Head/Tail shortcuts}
\usage{
first(x, ..., n = 1)

last(x, ..., n = 1)
}
\arguments{
\item{x}{vector object}

\item{...}{passed on to head or tail}

\item{n}{the new number to take of only one.}
}
\description{
Shortcuts for \code{head(x,1)} and \code{tail(x, 1)}
}
