\name{take_names}
\alias{dseval_left}
\alias{dseval_right}
\alias{get_vars}
\alias{take_names}
\title{Internal Commandes}
\usage{
  take_names(x)

  get_vars(x)

  dseval_left(x, y, idf)

  dseval_right(y, x, idf)
}
\arguments{
  \item{x}{parsed left side of formula}

  \item{y}{parsed right side of formula}

  \item{idf}{idata.frame object}
}
\description{
  Internal Commandes
}

