% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_call}
\alias{make_call}
\title{Make a call with extra arguments incorporated into call.}
\usage{
make_call(args, ..., what, quote = F, envir = parent.frame())
}
\arguments{
\item{args}{a list of arguments}

\item{...}{extra arguments to be incorporated into \code{args}}

\item{what}{the function to execute}

\item{quote}{should the arguments be quoted}

\item{envir}{the environment to call the function in}
}
\description{
Useful for using with \code{plyr} functions
}
\seealso{
\code{\link{do.call}} which this function wraps.
}
