% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.R
\name{pval}
\alias{pval}
\title{Extract a p-value from a test result.}
\usage{
pval(x, extended = F, ...)
}
\arguments{
\item{x}{a testing result object}

\item{extended}{should an extended result be given or a single p-value.}

\item{...}{extra arguments passed to methods.}
}
\value{
either a single value (\code{extended=FALSE}) representing the
p-value of the test or a single row.
\link{data.frame} object that also includes extra information such as
}
\description{
Extract a p-value from a test result.
}
\details{
This is a generic helper function for extracting p values from objects.
The idea being to extract the overall p-value for the model that can
be interpreted simply.
}
