% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/T.R
\name{.T}
\alias{.T}
\title{create a text vector}
\usage{
.T(...)
}
\arguments{
\item{...}{names, quoted or not, no substitution made}
}
\description{
create a text vector
}
\examples{
.T(min, mean, 'median')
}

