% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Table1}
\alias{Table1}
\title{Create a table of descriptives of a dataset.}
\usage{
Table1(formula, data)
}
\arguments{
\item{formula}{a formula description of the table.}

\item{data}{a data.frame or environment to extract the data from.}
}
\description{
Create a table of descriptives of a dataset.
}

