% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fill_v}
\alias{fill_v}
\title{Fill vector to length with a specified value}
\usage{
fill_v(x, l = length(x), with = last(x), after = length(x))
}
\arguments{
\item{x}{vector}

\item{l}{length}

\item{with}{What to fill with}

\item{after}{where to insert}
}
\description{
Fill vector to length with a specified value
}

