% Generated by roxygen2 (4.0.0): do not edit by hand
\name{take_names}
\alias{apply_by_two_fun}
\alias{dseval_fork}
\alias{dseval_left}
\alias{dseval_left_var}
\alias{dseval_right}
\alias{get_vars}
\alias{take_names}
\title{Internal formula Commands}
\usage{
take_names(x)

get_vars(x)

dseval_left(x, y, idf)

dseval_left_var(x, y, idf)

dseval_right(y, x, idf)

apply_by_two_fun(fun, x, y, idf)

dseval_fork(var, y, x, idf, gen_call)
}
\arguments{
\item{x}{parsed left side of formula}

\item{y}{parsed right side of formula}

\item{idf}{idata.frame object}

\item{fun}{function to apply}

\item{var}{variable to extract call from}

\item{gen_call}{generic call}
}
\description{
Internal formula Commands

Internal formula manipulation and evaluation
}

