% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wargs.R
\name{redirf}
\alias{redirf}
\title{Create a function that redirects to the named function.}
\usage{
redirf(f, envir = parent.frame())
}
\arguments{
\item{f}{a function to wrap a call around}

\item{envir}{environment to use for the function.}
}
\description{
This is usefull for debugging to know what function has been called
form within do.call or plyr functions.
}

