% Generated by roxygen2 (4.0.0): do not edit by hand
\name{seq_consecutive}
\alias{seq_consecutive}
\title{compute an indicator to group consecutive values}
\usage{
seq_consecutive(x, ...)
}
\arguments{
\item{x}{a vector}

\item{...}{ignored, might be used for forward compatibility.}
}
\value{
an integer vector.
}
\description{
computes a vector that changes every time the element is different from the
previous.
}

